Introduction
GitHub is a place for non-profits to share Blackbaud CRM SDK and API Code Samples. You are free to grab a local copy of one or more of the project repositories (such as the API Cookbook) and play with the samples. GitHub is based on git which is a free and open source distributed version control system designed to handle everything from small to very large projects with speed and efficiency. GitHub is an excellent platform for improving existing and adding new samples.